+++ packages/rx_bloc_cli/bin/compile.sh
@@ -9,6 +9,6 @@ dart run rx_bloc_cli create \
   --enable-feature-deeplinks=true\
   --enable-feature-widget-toolkit=true \
   --enable-social-logins=true \
-  --enable-internationalisation=true \
+  --enable-change-language=true \
   example/test_app
 cp example/test_app/README.md example/