+++ packages/rx_bloc_cli/bin/compile_bundles.sh
@@ -19,7 +19,7 @@ dart run mason_cli:mason bundle \
 dart run mason_cli:mason bundle \
   -t dart \
   -o lib/src/templates/ \
-  mason_templates/bricks/lib_internationalisation
+  mason_templates/bricks/lib_change_language
 
 dart run mason_cli:mason bundle \
   -t dart \