+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/common_blocs/coordinator_bloc.dart
@@ -5,8 +5,8 @@ import 'dart:async';
 import 'package:rx_bloc/rx_bloc.dart';
 import 'package:rxdart/rxdart.dart';
 
-{{#enable_internationalisation}}
-import 'package:widget_toolkit/language_picker.dart' show LanguageModel;{{/enable_internationalisation}}
+{{#enable_change_language}}
+import 'package:widget_toolkit/language_picker.dart' show LanguageModel;{{/enable_change_language}}
 
 import '../models/errors/error_model.dart';
 
@@ -20,15 +20,15 @@ abstract class CoordinatorEvents {
     required ErrorModel error,
     String? stackTrace,
   });
-  {{#enable_internationalisation}}
-  void setCurrentLanguage(LanguageModel model);{{/enable_internationalisation}}
+  {{#enable_change_language}}
+  void setCurrentLanguage(LanguageModel model);{{/enable_change_language}}
 }
 
 abstract class CoordinatorStates {
   @RxBlocIgnoreState()
   Stream<bool> get isAuthenticated;
-  {{#enable_internationalisation}}
-  Stream<LanguageModel> get currentLanguage;{{/enable_internationalisation}}
+  {{#enable_change_language}}
+  Stream<LanguageModel> get currentLanguage;{{/enable_change_language}}
 }
 
 /// The coordinator bloc manages the communication between blocs.
@@ -39,8 +39,8 @@ abstract class CoordinatorStates {
 class CoordinatorBloc extends $CoordinatorBloc {
   @override
   Stream<bool> get isAuthenticated => _$authenticatedEvent;
-  {{#enable_internationalisation}}
+  {{#enable_change_language}}
   @override
   Stream<LanguageModel> _mapToCurrentLanguageState() =>
-      _$setCurrentLanguageEvent;{{/enable_internationalisation}}
+      _$setCurrentLanguageEvent;{{/enable_change_language}}
 }