+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_profile/di/profile_page_with_dependencies.dart
@@ -2,10 +2,10 @@
 
 import 'package:flutter/widgets.dart';
 import 'package:flutter_rx_bloc/flutter_rx_bloc.dart';
-import 'package:provider/provider.dart';{{#enable_internationalisation}}
-import '../../lib_internationalisation/data_sources/language_local_data_source.dart';
-import '../../lib_internationalisation/data_storages/language_picker_shared_preferences_instance.dart';
-import '../../lib_internationalisation/services/language_service_example.dart';{{/enable_internationalisation}}
+import 'package:provider/provider.dart';{{#enable_change_language}}
+import '../../lib_change_language/data_sources/language_local_data_source.dart';
+import '../../lib_change_language/data_storages/language_picker_shared_preferences_instance.dart';
+import '../../lib_change_language/services/language_service_example.dart';{{/enable_change_language}}
 import '../blocs/profile_bloc.dart';
 import '../views/profile_page.dart';
 
@@ -14,15 +14,15 @@ class ProfilePageWithDependencies extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => MultiProvider(
-        providers: [{{#enable_internationalisation}}
+        providers: [{{#enable_change_language}}
           ..._dataStorages,
           ..._dataSources,
-          ..._services,{{/enable_internationalisation}}
+          ..._services,{{/enable_change_language}}
           ..._blocs,
         ],
         child: const ProfilePage(),
       );
-  {{#enable_internationalisation}}
+  {{#enable_change_language}}
   List<Provider> get _dataStorages => [
         Provider<LanguagePickerSharedPreferencesInstance>(
           create: (context) => LanguagePickerSharedPreferencesInstance(),
@@ -43,7 +43,7 @@ class ProfilePageWithDependencies extends StatelessWidget {
             localDataSource: context.read(),
           ),
         ),
-      ];{{/enable_internationalisation}}
+      ];{{/enable_change_language}}
 
   List<RxBlocProvider> get _blocs => [
         RxBlocProvider<ProfileBlocType>(